move model files from schemas
