+++ MuffinWeb/manage.py
@@ -1,16 +1,15 @@
 #! /usr/bin/env python
 
 import os
+
 from flask_compress import Compress
 from flask_cors import CORS
-
 from flask_script import Manager
 from flask_security import SQLAlchemyUserDatastore, Security
+
 from MuffinWeb.muffin.schemas.RolesModel import RolesModel
-from MuffinWeb.muffin.schemas.UsersModel import UsersModel
-
+from MuffinWeb.muffin.models import UsersModel
 from muffin import create_app, db
-
 
 app = create_app(os.getenv('MUFFIN_CONFIG', 'default'))
 CORS(app)
