+++ MuffinWeb/muffin/models/UsersModel.py
@@ -1,0 +1,29 @@
+from flask_security.core import UserMixin
+
+from MuffinWeb.muffin import db
+from MuffinWeb.muffin.models.BaseModel import BaseModel
+from MuffinWeb.muffin.models.RolesUsersModel import roles_users
+
+__author__ = 'ife'
+
+
+class UsersModel(BaseModel, UserMixin):
+    __tablename__ = 'auth_user'
+    email = db.Column(db.String(255), nullable=False, unique=True)
+    password = db.Column(db.String(255), nullable=False)
+    first_name = db.Column(db.String(255))
+    last_name = db.Column(db.String(255))
+    active = db.Column(db.Boolean())
+    confirmed_at = db.Column(db.DateTime())
+    last_login_at = db.Column(db.DateTime())
+    current_login_at = db.Column(db.DateTime())
+    # Why 45 characters for IP Address ?
+    # See http://stackoverflow.com/questions/166132/maximum-length-of-the-textual-representation-of-an-ipv6-address/166157#166157
+    last_login_ip = db.Column(db.String(45))
+    current_login_ip = db.Column(db.String(45))
+    login_count = db.Column(db.Integer)
+    roles = db.relationship('RolesModel', secondary=roles_users,
+                            backref=db.backref('users', lazy='dynamic'))
+
+    def __repr__(self):
+        return '<User %r>' % self.email
